Modified queries and some logic

+++ Backend/database/householdQueries.js
@@ -5,11 +5,10 @@ module.exports = {
     hostQuery: 'UPDATE users SET host = ?, role = ? WHERE id = ?',
     dataQuery: 'SELECT h.houseId, h.houseName, h.initBudget, h.balanceDate, hu.userId, hu.userName, hu.role FROM households h JOIN householdUsers hu ON h.houseId = hu.houseId WHERE h.houseId = ?;',
     ownershipQuery: 'SELECT houseId FROM households WHERE houseName=?',
-    deleteQuery: 'DELETE FROM households WHERE houseId=?',
+    deleteQuery: 'DELETE FROM households WHERE houseId=? AND userId=?',
     updateroleHu: 'UPDATE householdUsers SET role = ? WHERE userId = ?',
     getAllHousesQuery: 'SELECT h.houseId, h.houseName, h.userId, hu.userName, hu.role FROM households h JOIN householdUsers hu ON h.houseId = hu.houseId ORDER BY h.houseId;',
     updatehouseIdHu: 'UPDATE householdUsers SET houseId = ? WHERE userId = ?',
-    selectInmates: 'SELECT userId FROM householdUsers WHERE houseId=?',
     deleteHouseActions: 'DELETE FROM transactions WHERE houseId=?',
    
 };
